+++ compiler/src/main/java/info/teksol/mindcode/MessageLevel.java
@@ -8,6 +8,10 @@ public enum MessageLevel {
 
     private final String title;
 
+    public boolean strongerOrEqual(MessageLevel other) {
+        return ordinal() >= other.ordinal();
+    }
+
     MessageLevel() {
         title = name().substring(0, 1).toUpperCase() + name().substring(1).toLowerCase();
     }