+++ compiler/src/main/java/info/teksol/mindcode/grammar/MindcodeBaseVisitor.java
@@ -165,6 +165,13 @@ public class MindcodeBaseVisitor<T> extends AbstractParseTreeVisitor<T> implemen
 	 * <p>The default implementation returns the result of calling
 	 * {@link #visitChildren} on {@code ctx}.</p>
 	 */
+	@Override public T visitOrphan_doc_comment(MindcodeParser.Orphan_doc_commentContext ctx) { return visitChildren(ctx); }
+	/**
+	 * {@inheritDoc}
+	 *
+	 * <p>The default implementation returns the result of calling
+	 * {@link #visitChildren} on {@code ctx}.</p>
+	 */
 	@Override public T visitCompiler_directive(MindcodeParser.Compiler_directiveContext ctx) { return visitChildren(ctx); }
 	/**
 	 * {@inheritDoc}