+++ compiler/src/test/java/info/teksol/mindcode/ast/AstNodeBuilderTest.java
@@ -956,7 +956,7 @@ class AstNodeBuilderTest extends AbstractAstTest {
         assertEquals(
                 prettyPrint(
                         new Seq(null,
-                                new FunctionDeclaration(null, false, false, false,
+                                new FunctionDeclaration(null, null, false, false, false,
                                         "delay",
                                         List.of(),
                                         new Seq(null,
@@ -1017,7 +1017,7 @@ class AstNodeBuilderTest extends AbstractAstTest {
                 prettyPrint(
                         new Seq(null,
                                 new Seq(null,
-                                        new FunctionDeclaration(null, false, false, false,
+                                        new FunctionDeclaration(null, null, false, false, false,
                                                 "foo",
                                                 List.of(),
                                                 new Seq(null,
@@ -1048,7 +1048,7 @@ class AstNodeBuilderTest extends AbstractAstTest {
                 prettyPrint(
                         new Seq(null,
                                 new Seq(null,
-                                        new FunctionDeclaration(null, false, false, false,
+                                        new FunctionDeclaration(null, null, false, false, false,
                                                 "foo",
                                                 List.of(new FunctionParameter(null, "s", false, false, false)),
                                                 new Seq(null,
@@ -1077,7 +1077,7 @@ class AstNodeBuilderTest extends AbstractAstTest {
                 prettyPrint(
                         new Seq(null,
                                 new Seq(null,
-                                        new FunctionDeclaration(null, false, false, false,
+                                        new FunctionDeclaration(null, null, false, false, false,
                                                 "foo",
                                                 List.of(
                                                         new FunctionParameter(null, "s", false, false, false),
@@ -1295,6 +1295,36 @@ class AstNodeBuilderTest extends AbstractAstTest {
                           end;
                         end;
                         print("End");
+                        """
+                )
+        );
+    }
+
+    @Test
+    void compilesCodeDoc() {
+        assertEquals(
+                new Seq(null,
+                        new FunctionDeclaration(null, "the doc", false, false, false,
+                                "foo",
+                                List.of(
+                                        new FunctionParameter(null, "s", false, false, false),
+                                        new FunctionParameter(null, "r", false, false, false)
+                                ),
+                                new Seq(null,
+                                        new BinaryOp(null,
+                                                new BinaryOp(null, new VarRef(null, "s"), "+", new NumericLiteral(null, "1")),
+                                                "+",
+                                                new VarRef(null, "r")
+                                        )
+                                )
+                        )
+                ),
+
+                translateToAst("""
+                        /**the doc*/
+                        def foo(s, r)
+                          s + 1 + r;
+                        end;
                         """
                 )
         );